requirements.txt
pygame==1.9.6
PyOpenGL==3.1.5
